list_rc_log.py: Split functionality off into library
To allow for additional email-based reporting scripts, move the generic
functions off to lib/stable_email.py.

flake8 was used to find some unused and incorrectly used variables. Code
was formatted with black.
