+++ .flake8
@@ -1,2 +1,5 @@
 [flake8]
 max-line-length = 120
+extend-ignore =
+    # See https://github.com/PyCQA/pycodestyle/issues/373
+    E203,
